Check history-info argument and other fixes
Numerous IDs can be provided as argument to history-info
and invalid IDs are caught as errors.

Previously the action content map used a string for the action.
This means that arbitrary strings can technically be inserted as
an action. Changing this to an enum guarantees that only
valid actions are inserted as keys.